Move CalloutNativeCommands interface to its own types file and move codegenNativeCommands to each NativeComponent file
