+++ packages/components/Callout/src/index.ts
@@ -1,6 +1,5 @@
 export { calloutName } from './Callout.types';
 export type {
-  CalloutNativeCommands,
   DirectionalHint,
   DismissBehaviors,
   ICalloutProps,
@@ -10,4 +9,5 @@ export type {
   ICalloutType,
   RestoreFocusEvent,
 } from './Callout.types';
+export type { CalloutNativeCommands } from './CalloutNativeCommands.types';
 export { Callout } from './Callout';
